Do not fail when HRD functions are not available.
Because some HRD rig interface implementations are lacking features we
need, instead of failing fatally when  a needed feature is missing, we
now do not require the missing feature. This means that some rigs will
not work  correctly, for example  setting split,  PTT or mode  may not
work if no such feature is present in the HRD interface.

This came  about due to  the TenTec Omvi  VII HRD interface  having no
split, vfo  selection or,  PTT facilities. With  this change  at least
some basic functionallity is possible.

Not being able to set split  is potentially dangerous since WSJT-X may
set tones assumeing a split TX  frequency. Until the HRD CAT interface
has a mechanisim to signal the missing capability we will have to hope
that users  realise the deficiency  and don't attempt to  enable split
working on such a setup.