Disallow mode changes from decode dbl-clicks when transmitting.
The message generation code doen't allow  for a mode change on the fly
so this has been disabled.

Also and  decode double click  that would  change a message  or change
mode or  change a frquency  that gets  disallowed is ignored  i.e. the
outgoing message is not changed either.

Remove a  duplicate Tx  message in  decode window that  crept in  in a
prior change.