Better version identification.
An  ambiguity  existed  between  release  candidates  and  development
versions  as there  was no  definitive identification  for development
versions.

Added a  new version number  suffix of  '-devel' which appears  if the
release candidate number is missing  or zero in Versions.cmake and the
revision is not marked as a release version, also in Versions.cmake.

Also re-factored setting of version number suffiexes to a single place
in in CMake/VersionCompute.cmake.