Only poll split frequency if in split mode
Because PowerSDR for the FlexRadio rigs doesn't implement split or VFO
B,  querying the  TX VFO  in transceiver  polls causes  an error  with
them. This query is unecessary. Removing  it will cause a delay of one
poll period  after split is set  for the Tx frequency  to appear. This
should be  harmless, if not  then there  is a possibility  of querying
split before  the Tx frequency. This  seems obvious but I  believe the
order matters with some rigs so it needs more investigation.