WSJT-X User Guide location variable
In order to support the WSJT-X  super build in LOCAL mode the location
of the User Guide needs to be specified.

A CMake cache  variable WSJT_MANUAL_DIRECTORY has been  added to allow
the URI of the User Guide directory to be specified at configure time.
The default value remains as the project web site.