Add CMake option to exclude sample files from the embedded resources
The compilation of the WSJT-X resources file requires a lot of virtual
memory  because of  the  large .WAV  files  included.  For  developers
building on very limited resource  machines like single board embedded
platforms there is  now an option WSJT_EMBED_SAMPLES which  can be set
to OFF to skip the inclusion of the sample files in the WSJT_X binary.

The sample files may still be  added manually to the target machine if
required.